+++ administrator/components/com_crowdfundingfinance/CHANGELOG.md
@@ -1,7 +1,9 @@
 Crowdfunding Finance Changelog
 ===============================
+
 ###v2.4
 * Improved amount formatting.
+* It was made compatible with Crowdfunding v2.5.
 
 ###v2.3.1
 * Fixed an issue with Stripe in the plugin Crowdfunding - Payout Options.